refactor(ToastNotification): move List to Portal subcomponents
